fixed review update and home

+++ controllers/reviewController.js
@@ -64,7 +64,7 @@ exports.createReview = async (req, res, next) => {
 exports.updateReview = async (req, res, next) => {
   try {
     const { reviewId, placeId } = req.params;
-    const { title, review, star } = req.body;
+    const { title, review, star, review_pic } = req.body;
     const existReview = await Review.findOne({ where: { id: reviewId } });
     const place = await Place.findOne({ where: { id: placeId } });
     const reviewPic = await ReviewPic.findAll({
@@ -95,23 +95,44 @@ exports.updateReview = async (req, res, next) => {
       },
       { where: { id: reviewId } }
     );
+    const objReviewPic = JSON.parse(JSON.stringify(reviewPic, null, 2));
+    const picLength =
+      review_pic && req.files.review_pic
+        ? typeof review_pic === "string"
+          ? 1 + req.files.review_pic.length
+          : typeof review_pic === "object"
+          ? review_pic.length + req.files.review_pic.length
+          : req.files.review_pic.length
+        : typeof review_pic === "undefined"
+        ? req.files.review_pic.length
+        : typeof review_pic === "string"
+        ? 1
+        : review_pic.length;
 
     const lengthIdx =
-      reviewPic.length <= req.files.review_pic?.length
-        ? req.files.review_pic.length
-        : reviewPic.length;
+      objReviewPic.length <= picLength ? picLength : objReviewPic.length;
 
+    let i = 0;
     for (let idx = 0; idx < lengthIdx; idx++) {
-      if (req.files?.review_pic && req.files?.review_pic[idx]) {
-        if (reviewPic[idx]?.reviewPic) {
-          const splited = reviewPic[idx].reviewPic.split("/");
-          const publicId = splited[splited.length - 1].split(".")[0];
-          await cloudinary.destroy(publicId);
-        }
+      const pic =
+        typeof review_pic === "string"
+          ? review_pic
+          : typeof review_pic === "undefined"
+          ? null
+          : review_pic[idx];
 
-        const result = await cloudinary.upload(req.files.review_pic[idx].path);
-        if (reviewPic[idx]?.id) {
-          if (req.files.review_pic[idx]) {
+      if (req.files?.review_pic) {
+        const result = await cloudinary.upload(req.files.review_pic[i].path);
+        if (reviewPic[idx]?.reviewPic !== pic && pic !== undefined) {
+          if (reviewPic[idx]?.reviewPic) {
+            console.log("in");
+            const splited = reviewPic[idx].reviewPic.split("/");
+            const publicId = splited[splited.length - 1].split(".")[0];
+            await cloudinary.destroy(publicId);
+          }
+
+          i += 1;
+          if (reviewPic[idx]?.id) {
             await ReviewPic.update(
               {
                 reviewPic: result.secure_url,
@@ -121,18 +142,21 @@ exports.updateReview = async (req, res, next) => {
             );
           }
         } else {
-          await ReviewPic.create({
-            reviewPic: result.secure_url,
-            review_id: reviewId,
-          });
+          if (!reviewPic[idx]?.id) {
+            await ReviewPic.create({
+              reviewPic: result.secure_url,
+              review_id: reviewId,
+            });
+          }
         }
-      } else {
-        if (reviewPic[idx]?.reviewPic) {
+      } else if (review_pic) {
+        if (objReviewPic[idx].reviewPic !== pic) {
           const splited = reviewPic[idx].reviewPic.split("/");
           const publicId = splited[splited.length - 1].split(".")[0];
           await cloudinary.destroy(publicId);
+
+          await ReviewPic.destroy({ where: { id: reviewPic[idx].id } });
         }
-        await ReviewPic.destroy({ where: { id: reviewPic[idx].id } });
       }
     }
     calculateScore(place);
